updated location, conditional render
