+++ frontend/src/components/Navbar/Navbar.module.css
@@ -88,3 +88,24 @@
     font-size: 0.9rem;
   }
 }
+
+.navLogoWrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0.5rem;
+}
+
+.randomButton {
+  padding: 0.25rem 0.375rem;
+  font-size: 12px;
+  font-weight: 500;
+  font-size: 1rem;
+  margin-top: 0.25rem;
+  margin-left: -0.6rem;
+}
+
+.randomButton:disabled {
+  cursor: not-allowed;
+  color: gray;
+}
