+++ frontend/src/components/Navbar/Navbar.tsx
@@ -1,14 +1,15 @@
 import React, { useState } from 'react';
-import { NavLink, useLocation } from 'react-router-dom';
+import { NavLink, useLocation, useNavigate } from 'react-router-dom';
 import clsx from 'clsx';
-import { Nav, Navbar } from 'react-bootstrap';
+import { Nav, Navbar, Button } from 'react-bootstrap';
 import DarkModeButton from './DarkModeButton';
 import Logo from './Logo';
 import MeDropdown from './MeDropdown';
 import { API_ENDPOINT } from '../../config';
+import { useSearch } from '../../contexts/searchContext';
 import { logout } from '../../queries/api';
 import { useStore } from '../../store';
-import { scrollToTop } from '../../utilities/display';
+import { scrollToTop, createCourseModalLink } from '../../utilities/display';
 import LastUpdated from '../Search/LastUpdated';
 import { NavbarCatalogSearch } from '../Search/NavbarCatalogSearch';
 import { SurfaceComponent } from '../Typography';
@@ -56,9 +57,25 @@ export default function CourseTableNavbar() {
   const location = useLocation();
   const [navExpanded, setNavExpanded] = useState(false);
   const isMobile = useStore((state) => state.isMobile);
+  const { searchData } = useSearch();
+  const navigate = useNavigate();
+
+  const fetchRandomCourse = () => {
+    if (searchData && searchData.length > 0) {
+      const randomCourse =
+        searchData[Math.floor(Math.random() * searchData.length)];
+      const courseModalLink = createCourseModalLink(
+        randomCourse,
+        new URLSearchParams(),
+      );
+      navigate(courseModalLink);
+    }
+  };
 
   const showCatalogSearch = !isMobile && location.pathname === '/catalog';
   const showWorksheetSearch = !isMobile && location.pathname === '/worksheet';
+  // Show random even on mobile for now
+  const showRandom = location.pathname === '/catalog';
 
   return (
     <SurfaceComponent className={styles.container}>
@@ -72,12 +89,24 @@ export default function CourseTableNavbar() {
           showCatalogSearch && styles.catalogSearchNavbar,
         )}
       >
-        {/* Logo in top left */}
-        <Nav className={clsx(styles.navLogo, 'navbar-brand')}>
-          <NavLink to="/">
-            <Logo icon={false} />
-          </NavLink>
-        </Nav>
+        {/* Logo in top left and random underneath */}
+        <div className={styles.navLogoWrapper}>
+          <Nav className={clsx(styles.navLogo, 'navbar-brand')}>
+            <NavLink to="/">
+              <Logo icon={false} />
+            </NavLink>
+          </Nav>
+          {showRandom && (
+            <Button
+              className={clsx(styles.randomButton)} // Do we need clsx here
+              variant="outline-primary"
+              onClick={fetchRandomCourse}
+              disabled={!searchData || searchData.length === 0}
+            >
+              Random Course
+            </Button>
+          )}
+        </div>
         {showCatalogSearch && <NavbarCatalogSearch />}
         {showWorksheetSearch && <NavbarWorksheetSearch />}
